chore: move package into auth
