+++ src/main/java/com/bside/starterapi/application/AuthService.java
@@ -3,9 +3,9 @@ package com.bside.starterapi.application;
 import com.bside.starterapi.config.auth.JwtUtils;
 import com.bside.starterapi.config.auth.UserDetailsImpl;
 import com.bside.starterapi.domain.user.*;
-import com.bside.starterapi.presentation.AuthenticateUserRequest;
-import com.bside.starterapi.presentation.JwtResponse;
-import com.bside.starterapi.presentation.RegisterUserRequest;
+import com.bside.starterapi.presentation.auth.AuthenticateUserRequest;
+import com.bside.starterapi.presentation.auth.JwtResponse;
+import com.bside.starterapi.presentation.auth.RegisterUserRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
