+++ src/main/java/com/bside/starterapi/presentation/auth/AuthenticateUserRequest.java
@@ -1,4 +1,4 @@
-package com.bside.starterapi.presentation;
+package com.bside.starterapi.presentation.auth;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
